requirements.txt
# Alfred AI Dependencies

# Core web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.9
jinja2==3.1.3

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Vector database
qdrant-client>=1.10.1,<2.0.0

# Document processing
langchain==0.1.20
langchain-community==0.0.38
langchain-qdrant==0.1.3

# Configuration
python-dotenv==1.0.1

# Utilities
aiofiles==23.2.1
httpx==0.27.0

# Redis cache
redis==5.0.1
redis[hiredis]==5.0.1
